@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.7.2
+  sdk: ^3.8.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -64,6 +64,12 @@ dependencies:
 
   # Webview inside app
   webview_flutter: ^4.13.0
+
+  # Secured Storage
+  flutter_secure_storage: ^9.2.4
+
+  # Security
+  crypto: ^3.0.6
 
 dev_dependencies:
   flutter_test:
