@@ -1,6 +1,5 @@
-// dart format width=80
-// coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
+// coverage:ignore-file
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
@@ -14,10 +13,11 @@ part of 'game_model.dart';
 T _$identity<T>(T value) => value;
 
 /// @nodoc
-mixin _$GameModel {
+mixin _$GameModel implements DiagnosticableTreeMixin {
 
- String get gameId; String get name; int get recommended; String get briefDescription; String get description; String get requirements; String get headerImage; List<String>? get screenshots; Map<String, dynamic>? get price; List<String> get categoriesID;// Sale related fields
- bool? get isSale; double? get discountPercent; DateTime? get saleStartDate; DateTime? get saleEndDate;// Field for User only
+ String get gameId; String get publisherId; String get name; int get recommended; String get briefDescription; String get description; String get requirements; ResourceModel get headerImage; double get price; List<CategoryModel> get categories; List<ResourceModel>? get resources; DateTime get releaseDate;// Sale related fields
+ bool? get isSale; double? get discountPercent; DateTime? get saleStartDate; DateTime? get saleEndDate;// Download related fields
+ List<String>? get binaries; List<String>? get exes;// Field for User only
  bool get isOwned; bool get installed; bool get favorite; double? get playtimeHours;
 /// Create a copy of GameModel
 /// with the given fields replaced by the non-null parameter values.
@@ -28,19 +28,25 @@ $GameModelCopyWith<GameModel> get copyWith => _$GameModelCopyWithImpl<GameModel>
   /// Serializes this GameModel to a JSON map.
   Map<String, dynamic> toJson();
 
+@override
+void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+  properties
+    ..add(DiagnosticsProperty('type', 'GameModel'))
+    ..add(DiagnosticsProperty('gameId', gameId))..add(DiagnosticsProperty('publisherId', publisherId))..add(DiagnosticsProperty('name', name))..add(DiagnosticsProperty('recommended', recommended))..add(DiagnosticsProperty('briefDescription', briefDescription))..add(DiagnosticsProperty('description', description))..add(DiagnosticsProperty('requirements', requirements))..add(DiagnosticsProperty('headerImage', headerImage))..add(DiagnosticsProperty('price', price))..add(DiagnosticsProperty('categories', categories))..add(DiagnosticsProperty('resources', resources))..add(DiagnosticsProperty('releaseDate', releaseDate))..add(DiagnosticsProperty('isSale', isSale))..add(DiagnosticsProperty('discountPercent', discountPercent))..add(DiagnosticsProperty('saleStartDate', saleStartDate))..add(DiagnosticsProperty('saleEndDate', saleEndDate))..add(DiagnosticsProperty('binaries', binaries))..add(DiagnosticsProperty('exes', exes))..add(DiagnosticsProperty('isOwned', isOwned))..add(DiagnosticsProperty('installed', installed))..add(DiagnosticsProperty('favorite', favorite))..add(DiagnosticsProperty('playtimeHours', playtimeHours));
+}
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is GameModel&&(identical(other.gameId, gameId) || other.gameId == gameId)&&(identical(other.name, name) || other.name == name)&&(identical(other.recommended, recommended) || other.recommended == recommended)&&(identical(other.briefDescription, briefDescription) || other.briefDescription == briefDescription)&&(identical(other.description, description) || other.description == description)&&(identical(other.requirements, requirements) || other.requirements == requirements)&&(identical(other.headerImage, headerImage) || other.headerImage == headerImage)&&const DeepCollectionEquality().equals(other.screenshots, screenshots)&&const DeepCollectionEquality().equals(other.price, price)&&const DeepCollectionEquality().equals(other.categoriesID, categoriesID)&&(identical(other.isSale, isSale) || other.isSale == isSale)&&(identical(other.discountPercent, discountPercent) || other.discountPercent == discountPercent)&&(identical(other.saleStartDate, saleStartDate) || other.saleStartDate == saleStartDate)&&(identical(other.saleEndDate, saleEndDate) || other.saleEndDate == saleEndDate)&&(identical(other.isOwned, isOwned) || other.isOwned == isOwned)&&(identical(other.installed, installed) || other.installed == installed)&&(identical(other.favorite, favorite) || other.favorite == favorite)&&(identical(other.playtimeHours, playtimeHours) || other.playtimeHours == playtimeHours));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is GameModel&&(identical(other.gameId, gameId) || other.gameId == gameId)&&(identical(other.publisherId, publisherId) || other.publisherId == publisherId)&&(identical(other.name, name) || other.name == name)&&(identical(other.recommended, recommended) || other.recommended == recommended)&&(identical(other.briefDescription, briefDescription) || other.briefDescription == briefDescription)&&(identical(other.description, description) || other.description == description)&&(identical(other.requirements, requirements) || other.requirements == requirements)&&(identical(other.headerImage, headerImage) || other.headerImage == headerImage)&&(identical(other.price, price) || other.price == price)&&const DeepCollectionEquality().equals(other.categories, categories)&&const DeepCollectionEquality().equals(other.resources, resources)&&(identical(other.releaseDate, releaseDate) || other.releaseDate == releaseDate)&&(identical(other.isSale, isSale) || other.isSale == isSale)&&(identical(other.discountPercent, discountPercent) || other.discountPercent == discountPercent)&&(identical(other.saleStartDate, saleStartDate) || other.saleStartDate == saleStartDate)&&(identical(other.saleEndDate, saleEndDate) || other.saleEndDate == saleEndDate)&&const DeepCollectionEquality().equals(other.binaries, binaries)&&const DeepCollectionEquality().equals(other.exes, exes)&&(identical(other.isOwned, isOwned) || other.isOwned == isOwned)&&(identical(other.installed, installed) || other.installed == installed)&&(identical(other.favorite, favorite) || other.favorite == favorite)&&(identical(other.playtimeHours, playtimeHours) || other.playtimeHours == playtimeHours));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,gameId,name,recommended,briefDescription,description,requirements,headerImage,const DeepCollectionEquality().hash(screenshots),const DeepCollectionEquality().hash(price),const DeepCollectionEquality().hash(categoriesID),isSale,discountPercent,saleStartDate,saleEndDate,isOwned,installed,favorite,playtimeHours);
+int get hashCode => Object.hashAll([runtimeType,gameId,publisherId,name,recommended,briefDescription,description,requirements,headerImage,price,const DeepCollectionEquality().hash(categories),const DeepCollectionEquality().hash(resources),releaseDate,isSale,discountPercent,saleStartDate,saleEndDate,const DeepCollectionEquality().hash(binaries),const DeepCollectionEquality().hash(exes),isOwned,installed,favorite,playtimeHours]);
 
 @override
-String toString() {
-  return 'GameModel(appId: $gameId, name: $name, recommended: $recommended, briefDescription: $briefDescription, description: $description, requirements: $requirements, headerImage: $headerImage, screenshots: $screenshots, price: $price, categoriesID: $categoriesID, isSale: $isSale, discountPercent: $discountPercent, saleStartDate: $saleStartDate, saleEndDate: $saleEndDate, isOwned: $isOwned, installed: $installed, favorite: $favorite, playtimeHours: $playtimeHours)';
+String toString({ DiagnosticLevel minLevel = DiagnosticLevel.info }) {
+  return 'GameModel(gameId: $gameId, publisherId: $publisherId, name: $name, recommended: $recommended, briefDescription: $briefDescription, description: $description, requirements: $requirements, headerImage: $headerImage, price: $price, categories: $categories, resources: $resources, releaseDate: $releaseDate, isSale: $isSale, discountPercent: $discountPercent, saleStartDate: $saleStartDate, saleEndDate: $saleEndDate, binaries: $binaries, exes: $exes, isOwned: $isOwned, installed: $installed, favorite: $favorite, playtimeHours: $playtimeHours)';
 }
 
 
@@ -51,11 +57,11 @@ abstract mixin class $GameModelCopyWith<$Res>  {
   factory $GameModelCopyWith(GameModel value, $Res Function(GameModel) _then) = _$GameModelCopyWithImpl;
 @useResult
 $Res call({
- String appId, String name, int recommended, String briefDescription, String description, String requirements, String headerImage, List<String>? screenshots, Map<String, dynamic>? price, List<String> categoriesID, bool? isSale, double? discountPercent, DateTime? saleStartDate, DateTime? saleEndDate, bool isOwned, bool installed, bool favorite, double? playtimeHours
+ String gameId, String publisherId, String name, int recommended, String briefDescription, String description, String requirements, ResourceModel headerImage, double price, List<CategoryModel> categories, List<ResourceModel>? resources, DateTime releaseDate, bool? isSale, double? discountPercent, DateTime? saleStartDate, DateTime? saleEndDate, List<String>? binaries, List<String>? exes, bool isOwned, bool installed, bool favorite, double? playtimeHours
 });
 
 
-
+$ResourceModelCopyWith<$Res> get headerImage;
 
 }
 /// @nodoc
@@ -68,77 +74,234 @@ class _$GameModelCopyWithImpl<$Res>
 
 /// Create a copy of GameModel
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? appId = null,Object? name = null,Object? recommended = null,Object? briefDescription = null,Object? description = null,Object? requirements = null,Object? headerImage = null,Object? screenshots = freezed,Object? price = freezed,Object? categoriesID = null,Object? isSale = freezed,Object? discountPercent = freezed,Object? saleStartDate = freezed,Object? saleEndDate = freezed,Object? isOwned = null,Object? installed = null,Object? favorite = null,Object? playtimeHours = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? gameId = null,Object? publisherId = null,Object? name = null,Object? recommended = null,Object? briefDescription = null,Object? description = null,Object? requirements = null,Object? headerImage = null,Object? price = null,Object? categories = null,Object? resources = freezed,Object? releaseDate = null,Object? isSale = freezed,Object? discountPercent = freezed,Object? saleStartDate = freezed,Object? saleEndDate = freezed,Object? binaries = freezed,Object? exes = freezed,Object? isOwned = null,Object? installed = null,Object? favorite = null,Object? playtimeHours = freezed,}) {
   return _then(_self.copyWith(
-appId: null == appId ? _self.gameId : appId // ignore: cast_nullable_to_non_nullable
+gameId: null == gameId ? _self.gameId : gameId // ignore: cast_nullable_to_non_nullable
+as String,publisherId: null == publisherId ? _self.publisherId : publisherId // ignore: cast_nullable_to_non_nullable
 as String,name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
 as String,recommended: null == recommended ? _self.recommended : recommended // ignore: cast_nullable_to_non_nullable
 as int,briefDescription: null == briefDescription ? _self.briefDescription : briefDescription // ignore: cast_nullable_to_non_nullable
 as String,description: null == description ? _self.description : description // ignore: cast_nullable_to_non_nullable
 as String,requirements: null == requirements ? _self.requirements : requirements // ignore: cast_nullable_to_non_nullable
 as String,headerImage: null == headerImage ? _self.headerImage : headerImage // ignore: cast_nullable_to_non_nullable
-as String,screenshots: freezed == screenshots ? _self.screenshots : screenshots // ignore: cast_nullable_to_non_nullable
-as List<String>?,price: freezed == price ? _self.price : price // ignore: cast_nullable_to_non_nullable
-as Map<String, dynamic>?,categoriesID: null == categoriesID ? _self.categoriesID : categoriesID // ignore: cast_nullable_to_non_nullable
-as List<String>,isSale: freezed == isSale ? _self.isSale : isSale // ignore: cast_nullable_to_non_nullable
+as ResourceModel,price: null == price ? _self.price : price // ignore: cast_nullable_to_non_nullable
+as double,categories: null == categories ? _self.categories : categories // ignore: cast_nullable_to_non_nullable
+as List<CategoryModel>,resources: freezed == resources ? _self.resources : resources // ignore: cast_nullable_to_non_nullable
+as List<ResourceModel>?,releaseDate: null == releaseDate ? _self.releaseDate : releaseDate // ignore: cast_nullable_to_non_nullable
+as DateTime,isSale: freezed == isSale ? _self.isSale : isSale // ignore: cast_nullable_to_non_nullable
 as bool?,discountPercent: freezed == discountPercent ? _self.discountPercent : discountPercent // ignore: cast_nullable_to_non_nullable
 as double?,saleStartDate: freezed == saleStartDate ? _self.saleStartDate : saleStartDate // ignore: cast_nullable_to_non_nullable
 as DateTime?,saleEndDate: freezed == saleEndDate ? _self.saleEndDate : saleEndDate // ignore: cast_nullable_to_non_nullable
-as DateTime?,isOwned: null == isOwned ? _self.isOwned : isOwned // ignore: cast_nullable_to_non_nullable
+as DateTime?,binaries: freezed == binaries ? _self.binaries : binaries // ignore: cast_nullable_to_non_nullable
+as List<String>?,exes: freezed == exes ? _self.exes : exes // ignore: cast_nullable_to_non_nullable
+as List<String>?,isOwned: null == isOwned ? _self.isOwned : isOwned // ignore: cast_nullable_to_non_nullable
 as bool,installed: null == installed ? _self.installed : installed // ignore: cast_nullable_to_non_nullable
 as bool,favorite: null == favorite ? _self.favorite : favorite // ignore: cast_nullable_to_non_nullable
 as bool,playtimeHours: freezed == playtimeHours ? _self.playtimeHours : playtimeHours // ignore: cast_nullable_to_non_nullable
 as double?,
   ));
 }
-
+/// Create a copy of GameModel
+/// with the given fields replaced by the non-null parameter values.
+@override
+@pragma('vm:prefer-inline')
+$ResourceModelCopyWith<$Res> get headerImage {
+  
+  return $ResourceModelCopyWith<$Res>(_self.headerImage, (value) {
+    return _then(_self.copyWith(headerImage: value));
+  });
+}
 }
 
+
+/// Adds pattern-matching-related methods to [GameModel].
+extension GameModelPatterns on GameModel {
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _GameModel value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _GameModel() when $default != null:
+return $default(_that);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _GameModel value)  $default,){
+final _that = this;
+switch (_that) {
+case _GameModel():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _GameModel value)?  $default,){
+final _that = this;
+switch (_that) {
+case _GameModel() when $default != null:
+return $default(_that);case _:
+  return null;
+
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String gameId,  String publisherId,  String name,  int recommended,  String briefDescription,  String description,  String requirements,  ResourceModel headerImage,  double price,  List<CategoryModel> categories,  List<ResourceModel>? resources,  DateTime releaseDate,  bool? isSale,  double? discountPercent,  DateTime? saleStartDate,  DateTime? saleEndDate,  List<String>? binaries,  List<String>? exes,  bool isOwned,  bool installed,  bool favorite,  double? playtimeHours)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _GameModel() when $default != null:
+return $default(_that.gameId,_that.publisherId,_that.name,_that.recommended,_that.briefDescription,_that.description,_that.requirements,_that.headerImage,_that.price,_that.categories,_that.resources,_that.releaseDate,_that.isSale,_that.discountPercent,_that.saleStartDate,_that.saleEndDate,_that.binaries,_that.exes,_that.isOwned,_that.installed,_that.favorite,_that.playtimeHours);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String gameId,  String publisherId,  String name,  int recommended,  String briefDescription,  String description,  String requirements,  ResourceModel headerImage,  double price,  List<CategoryModel> categories,  List<ResourceModel>? resources,  DateTime releaseDate,  bool? isSale,  double? discountPercent,  DateTime? saleStartDate,  DateTime? saleEndDate,  List<String>? binaries,  List<String>? exes,  bool isOwned,  bool installed,  bool favorite,  double? playtimeHours)  $default,) {final _that = this;
+switch (_that) {
+case _GameModel():
+return $default(_that.gameId,_that.publisherId,_that.name,_that.recommended,_that.briefDescription,_that.description,_that.requirements,_that.headerImage,_that.price,_that.categories,_that.resources,_that.releaseDate,_that.isSale,_that.discountPercent,_that.saleStartDate,_that.saleEndDate,_that.binaries,_that.exes,_that.isOwned,_that.installed,_that.favorite,_that.playtimeHours);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String gameId,  String publisherId,  String name,  int recommended,  String briefDescription,  String description,  String requirements,  ResourceModel headerImage,  double price,  List<CategoryModel> categories,  List<ResourceModel>? resources,  DateTime releaseDate,  bool? isSale,  double? discountPercent,  DateTime? saleStartDate,  DateTime? saleEndDate,  List<String>? binaries,  List<String>? exes,  bool isOwned,  bool installed,  bool favorite,  double? playtimeHours)?  $default,) {final _that = this;
+switch (_that) {
+case _GameModel() when $default != null:
+return $default(_that.gameId,_that.publisherId,_that.name,_that.recommended,_that.briefDescription,_that.description,_that.requirements,_that.headerImage,_that.price,_that.categories,_that.resources,_that.releaseDate,_that.isSale,_that.discountPercent,_that.saleStartDate,_that.saleEndDate,_that.binaries,_that.exes,_that.isOwned,_that.installed,_that.favorite,_that.playtimeHours);case _:
+  return null;
+
+}
+}
+
+}
 
 /// @nodoc
 @JsonSerializable()
 
-class _GameModel implements GameModel {
-  const _GameModel({required this.gameId, required this.name, required this.recommended, required this.briefDescription, required this.description, required this.requirements, required this.headerImage, final  List<String>? screenshots, final  Map<String, dynamic>? price, required final  List<String> categoriesID, this.isSale, this.discountPercent, this.saleStartDate, this.saleEndDate, this.isOwned = false, this.installed = false, this.favorite = false, this.playtimeHours}): _screenshots = screenshots,_price = price,_categoriesID = categoriesID;
+class _GameModel with DiagnosticableTreeMixin implements GameModel {
+  const _GameModel({required this.gameId, required this.publisherId, required this.name, required this.recommended, required this.briefDescription, required this.description, required this.requirements, required this.headerImage, required this.price, required final  List<CategoryModel> categories, required final  List<ResourceModel>? resources, required this.releaseDate, this.isSale, this.discountPercent, this.saleStartDate, this.saleEndDate, final  List<String>? binaries, final  List<String>? exes, this.isOwned = false, this.installed = false, this.favorite = false, this.playtimeHours}): _categories = categories,_resources = resources,_binaries = binaries,_exes = exes;
   factory _GameModel.fromJson(Map<String, dynamic> json) => _$GameModelFromJson(json);
 
 @override final  String gameId;
+@override final  String publisherId;
 @override final  String name;
 @override final  int recommended;
 @override final  String briefDescription;
 @override final  String description;
 @override final  String requirements;
-@override final  String headerImage;
- final  List<String>? _screenshots;
-@override List<String>? get screenshots {
-  final value = _screenshots;
+@override final  ResourceModel headerImage;
+@override final  double price;
+ final  List<CategoryModel> _categories;
+@override List<CategoryModel> get categories {
+  if (_categories is EqualUnmodifiableListView) return _categories;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableListView(_categories);
+}
+
+ final  List<ResourceModel>? _resources;
+@override List<ResourceModel>? get resources {
+  final value = _resources;
   if (value == null) return null;
-  if (_screenshots is EqualUnmodifiableListView) return _screenshots;
+  if (_resources is EqualUnmodifiableListView) return _resources;
   // ignore: implicit_dynamic_type
   return EqualUnmodifiableListView(value);
 }
 
- final  Map<String, dynamic>? _price;
-@override Map<String, dynamic>? get price {
-  final value = _price;
-  if (value == null) return null;
-  if (_price is EqualUnmodifiableMapView) return _price;
-  // ignore: implicit_dynamic_type
-  return EqualUnmodifiableMapView(value);
-}
-
- final  List<String> _categoriesID;
-@override List<String> get categoriesID {
-  if (_categoriesID is EqualUnmodifiableListView) return _categoriesID;
-  // ignore: implicit_dynamic_type
-  return EqualUnmodifiableListView(_categoriesID);
-}
-
+@override final  DateTime releaseDate;
 // Sale related fields
 @override final  bool? isSale;
 @override final  double? discountPercent;
 @override final  DateTime? saleStartDate;
 @override final  DateTime? saleEndDate;
+// Download related fields
+ final  List<String>? _binaries;
+// Download related fields
+@override List<String>? get binaries {
+  final value = _binaries;
+  if (value == null) return null;
+  if (_binaries is EqualUnmodifiableListView) return _binaries;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableListView(value);
+}
+
+ final  List<String>? _exes;
+@override List<String>? get exes {
+  final value = _exes;
+  if (value == null) return null;
+  if (_exes is EqualUnmodifiableListView) return _exes;
+  // ignore: implicit_dynamic_type
+  return EqualUnmodifiableListView(value);
+}
+
 // Field for User only
 @override@JsonKey() final  bool isOwned;
 @override@JsonKey() final  bool installed;
@@ -155,19 +318,25 @@ _$GameModelCopyWith<_GameModel> get copyWith => __$GameModelCopyWithImpl<_GameMo
 Map<String, dynamic> toJson() {
   return _$GameModelToJson(this, );
 }
+@override
+void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+  properties
+    ..add(DiagnosticsProperty('type', 'GameModel'))
+    ..add(DiagnosticsProperty('gameId', gameId))..add(DiagnosticsProperty('publisherId', publisherId))..add(DiagnosticsProperty('name', name))..add(DiagnosticsProperty('recommended', recommended))..add(DiagnosticsProperty('briefDescription', briefDescription))..add(DiagnosticsProperty('description', description))..add(DiagnosticsProperty('requirements', requirements))..add(DiagnosticsProperty('headerImage', headerImage))..add(DiagnosticsProperty('price', price))..add(DiagnosticsProperty('categories', categories))..add(DiagnosticsProperty('resources', resources))..add(DiagnosticsProperty('releaseDate', releaseDate))..add(DiagnosticsProperty('isSale', isSale))..add(DiagnosticsProperty('discountPercent', discountPercent))..add(DiagnosticsProperty('saleStartDate', saleStartDate))..add(DiagnosticsProperty('saleEndDate', saleEndDate))..add(DiagnosticsProperty('binaries', binaries))..add(DiagnosticsProperty('exes', exes))..add(DiagnosticsProperty('isOwned', isOwned))..add(DiagnosticsProperty('installed', installed))..add(DiagnosticsProperty('favorite', favorite))..add(DiagnosticsProperty('playtimeHours', playtimeHours));
+}
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _GameModel&&(identical(other.gameId, gameId) || other.gameId == gameId)&&(identical(other.name, name) || other.name == name)&&(identical(other.recommended, recommended) || other.recommended == recommended)&&(identical(other.briefDescription, briefDescription) || other.briefDescription == briefDescription)&&(identical(other.description, description) || other.description == description)&&(identical(other.requirements, requirements) || other.requirements == requirements)&&(identical(other.headerImage, headerImage) || other.headerImage == headerImage)&&const DeepCollectionEquality().equals(other._screenshots, _screenshots)&&const DeepCollectionEquality().equals(other._price, _price)&&const DeepCollectionEquality().equals(other._categoriesID, _categoriesID)&&(identical(other.isSale, isSale) || other.isSale == isSale)&&(identical(other.discountPercent, discountPercent) || other.discountPercent == discountPercent)&&(identical(other.saleStartDate, saleStartDate) || other.saleStartDate == saleStartDate)&&(identical(other.saleEndDate, saleEndDate) || other.saleEndDate == saleEndDate)&&(identical(other.isOwned, isOwned) || other.isOwned == isOwned)&&(identical(other.installed, installed) || other.installed == installed)&&(identical(other.favorite, favorite) || other.favorite == favorite)&&(identical(other.playtimeHours, playtimeHours) || other.playtimeHours == playtimeHours));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _GameModel&&(identical(other.gameId, gameId) || other.gameId == gameId)&&(identical(other.publisherId, publisherId) || other.publisherId == publisherId)&&(identical(other.name, name) || other.name == name)&&(identical(other.recommended, recommended) || other.recommended == recommended)&&(identical(other.briefDescription, briefDescription) || other.briefDescription == briefDescription)&&(identical(other.description, description) || other.description == description)&&(identical(other.requirements, requirements) || other.requirements == requirements)&&(identical(other.headerImage, headerImage) || other.headerImage == headerImage)&&(identical(other.price, price) || other.price == price)&&const DeepCollectionEquality().equals(other._categories, _categories)&&const DeepCollectionEquality().equals(other._resources, _resources)&&(identical(other.releaseDate, releaseDate) || other.releaseDate == releaseDate)&&(identical(other.isSale, isSale) || other.isSale == isSale)&&(identical(other.discountPercent, discountPercent) || other.discountPercent == discountPercent)&&(identical(other.saleStartDate, saleStartDate) || other.saleStartDate == saleStartDate)&&(identical(other.saleEndDate, saleEndDate) || other.saleEndDate == saleEndDate)&&const DeepCollectionEquality().equals(other._binaries, _binaries)&&const DeepCollectionEquality().equals(other._exes, _exes)&&(identical(other.isOwned, isOwned) || other.isOwned == isOwned)&&(identical(other.installed, installed) || other.installed == installed)&&(identical(other.favorite, favorite) || other.favorite == favorite)&&(identical(other.playtimeHours, playtimeHours) || other.playtimeHours == playtimeHours));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,gameId,name,recommended,briefDescription,description,requirements,headerImage,const DeepCollectionEquality().hash(_screenshots),const DeepCollectionEquality().hash(_price),const DeepCollectionEquality().hash(_categoriesID),isSale,discountPercent,saleStartDate,saleEndDate,isOwned,installed,favorite,playtimeHours);
+int get hashCode => Object.hashAll([runtimeType,gameId,publisherId,name,recommended,briefDescription,description,requirements,headerImage,price,const DeepCollectionEquality().hash(_categories),const DeepCollectionEquality().hash(_resources),releaseDate,isSale,discountPercent,saleStartDate,saleEndDate,const DeepCollectionEquality().hash(_binaries),const DeepCollectionEquality().hash(_exes),isOwned,installed,favorite,playtimeHours]);
 
 @override
-String toString() {
-  return 'GameModel(appId: $gameId, name: $name, recommended: $recommended, briefDescription: $briefDescription, description: $description, requirements: $requirements, headerImage: $headerImage, screenshots: $screenshots, price: $price, categoriesID: $categoriesID, isSale: $isSale, discountPercent: $discountPercent, saleStartDate: $saleStartDate, saleEndDate: $saleEndDate, isOwned: $isOwned, installed: $installed, favorite: $favorite, playtimeHours: $playtimeHours)';
+String toString({ DiagnosticLevel minLevel = DiagnosticLevel.info }) {
+  return 'GameModel(gameId: $gameId, publisherId: $publisherId, name: $name, recommended: $recommended, briefDescription: $briefDescription, description: $description, requirements: $requirements, headerImage: $headerImage, price: $price, categories: $categories, resources: $resources, releaseDate: $releaseDate, isSale: $isSale, discountPercent: $discountPercent, saleStartDate: $saleStartDate, saleEndDate: $saleEndDate, binaries: $binaries, exes: $exes, isOwned: $isOwned, installed: $installed, favorite: $favorite, playtimeHours: $playtimeHours)';
 }
 
 
@@ -178,11 +347,11 @@ abstract mixin class _$GameModelCopyWith<$Res> implements $GameModelCopyWith<$Re
   factory _$GameModelCopyWith(_GameModel value, $Res Function(_GameModel) _then) = __$GameModelCopyWithImpl;
 @override @useResult
 $Res call({
- String appId, String name, int recommended, String briefDescription, String description, String requirements, String headerImage, List<String>? screenshots, Map<String, dynamic>? price, List<String> categoriesID, bool? isSale, double? discountPercent, DateTime? saleStartDate, DateTime? saleEndDate, bool isOwned, bool installed, bool favorite, double? playtimeHours
+ String gameId, String publisherId, String name, int recommended, String briefDescription, String description, String requirements, ResourceModel headerImage, double price, List<CategoryModel> categories, List<ResourceModel>? resources, DateTime releaseDate, bool? isSale, double? discountPercent, DateTime? saleStartDate, DateTime? saleEndDate, List<String>? binaries, List<String>? exes, bool isOwned, bool installed, bool favorite, double? playtimeHours
 });
 
 
-
+@override $ResourceModelCopyWith<$Res> get headerImage;
 
 }
 /// @nodoc
@@ -195,23 +364,27 @@ class __$GameModelCopyWithImpl<$Res>
 
 /// Create a copy of GameModel
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? appId = null,Object? name = null,Object? recommended = null,Object? briefDescription = null,Object? description = null,Object? requirements = null,Object? headerImage = null,Object? screenshots = freezed,Object? price = freezed,Object? categoriesID = null,Object? isSale = freezed,Object? discountPercent = freezed,Object? saleStartDate = freezed,Object? saleEndDate = freezed,Object? isOwned = null,Object? installed = null,Object? favorite = null,Object? playtimeHours = freezed,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? gameId = null,Object? publisherId = null,Object? name = null,Object? recommended = null,Object? briefDescription = null,Object? description = null,Object? requirements = null,Object? headerImage = null,Object? price = null,Object? categories = null,Object? resources = freezed,Object? releaseDate = null,Object? isSale = freezed,Object? discountPercent = freezed,Object? saleStartDate = freezed,Object? saleEndDate = freezed,Object? binaries = freezed,Object? exes = freezed,Object? isOwned = null,Object? installed = null,Object? favorite = null,Object? playtimeHours = freezed,}) {
   return _then(_GameModel(
-gameId: null == appId ? _self.gameId : appId // ignore: cast_nullable_to_non_nullable
+gameId: null == gameId ? _self.gameId : gameId // ignore: cast_nullable_to_non_nullable
+as String,publisherId: null == publisherId ? _self.publisherId : publisherId // ignore: cast_nullable_to_non_nullable
 as String,name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
 as String,recommended: null == recommended ? _self.recommended : recommended // ignore: cast_nullable_to_non_nullable
 as int,briefDescription: null == briefDescription ? _self.briefDescription : briefDescription // ignore: cast_nullable_to_non_nullable
 as String,description: null == description ? _self.description : description // ignore: cast_nullable_to_non_nullable
 as String,requirements: null == requirements ? _self.requirements : requirements // ignore: cast_nullable_to_non_nullable
 as String,headerImage: null == headerImage ? _self.headerImage : headerImage // ignore: cast_nullable_to_non_nullable
-as String,screenshots: freezed == screenshots ? _self._screenshots : screenshots // ignore: cast_nullable_to_non_nullable
-as List<String>?,price: freezed == price ? _self._price : price // ignore: cast_nullable_to_non_nullable
-as Map<String, dynamic>?,categoriesID: null == categoriesID ? _self._categoriesID : categoriesID // ignore: cast_nullable_to_non_nullable
-as List<String>,isSale: freezed == isSale ? _self.isSale : isSale // ignore: cast_nullable_to_non_nullable
+as ResourceModel,price: null == price ? _self.price : price // ignore: cast_nullable_to_non_nullable
+as double,categories: null == categories ? _self._categories : categories // ignore: cast_nullable_to_non_nullable
+as List<CategoryModel>,resources: freezed == resources ? _self._resources : resources // ignore: cast_nullable_to_non_nullable
+as List<ResourceModel>?,releaseDate: null == releaseDate ? _self.releaseDate : releaseDate // ignore: cast_nullable_to_non_nullable
+as DateTime,isSale: freezed == isSale ? _self.isSale : isSale // ignore: cast_nullable_to_non_nullable
 as bool?,discountPercent: freezed == discountPercent ? _self.discountPercent : discountPercent // ignore: cast_nullable_to_non_nullable
 as double?,saleStartDate: freezed == saleStartDate ? _self.saleStartDate : saleStartDate // ignore: cast_nullable_to_non_nullable
 as DateTime?,saleEndDate: freezed == saleEndDate ? _self.saleEndDate : saleEndDate // ignore: cast_nullable_to_non_nullable
-as DateTime?,isOwned: null == isOwned ? _self.isOwned : isOwned // ignore: cast_nullable_to_non_nullable
+as DateTime?,binaries: freezed == binaries ? _self._binaries : binaries // ignore: cast_nullable_to_non_nullable
+as List<String>?,exes: freezed == exes ? _self._exes : exes // ignore: cast_nullable_to_non_nullable
+as List<String>?,isOwned: null == isOwned ? _self.isOwned : isOwned // ignore: cast_nullable_to_non_nullable
 as bool,installed: null == installed ? _self.installed : installed // ignore: cast_nullable_to_non_nullable
 as bool,favorite: null == favorite ? _self.favorite : favorite // ignore: cast_nullable_to_non_nullable
 as bool,playtimeHours: freezed == playtimeHours ? _self.playtimeHours : playtimeHours // ignore: cast_nullable_to_non_nullable
@@ -219,7 +392,16 @@ as double?,
   ));
 }
 
-
+/// Create a copy of GameModel
+/// with the given fields replaced by the non-null parameter values.
+@override
+@pragma('vm:prefer-inline')
+$ResourceModelCopyWith<$Res> get headerImage {
+  
+  return $ResourceModelCopyWith<$Res>(_self.headerImage, (value) {
+    return _then(_self.copyWith(headerImage: value));
+  });
+}
 }
 
 // dart format on
