@@ -1,89 +1,92 @@
 import 'dart:convert';
 import 'package:flutter/foundation.dart';
 import 'package:http/http.dart' as http;
-import 'package:gameverse/domain/models/user_model/user_model.dart';
+import "package:crypto/crypto.dart";
+
+import 'package:gameverse/config/api_endpoints.dart';
+import 'package:gameverse/utils/response.dart';
 
 class AuthApiClient {
   final http.Client _client;
-  final String baseUrl;
   
-  AuthApiClient({
-    http.Client? client,
-    required this.baseUrl,
-  }) : _client = client ?? http.Client();
+  AuthApiClient({http.Client? client}) : _client = client ?? http.Client();
 
-  // Login with email and password
-  Future<UserModel?> login(String email, String password) async {
-    try {
-      final response = await _client.post(
-        Uri.parse('$baseUrl/auth/login'),
-        headers: {'Content-Type': 'application/json'},
-        body: json.encode({
-          'email': email,
-          'password': password,
-        }),
-      );
-      
-      if (response.statusCode == 200) {
-        final jsonData = json.decode(response.body);
-        return UserModel.fromJson(jsonData['user']);
-      } else {
-        debugPrint('Login failed: ${response.statusCode}');
-        return null;
-      }
-    } catch (e) {
-      debugPrint('Login error: $e');
-      return null;
-    }
+  Future<Response> login(String email, String password) async {
+    final bytePassword = utf8.encode(password);
+    final hashPassword = sha256.convert(bytePassword).toString();
+
+    final raw = await http.post(
+      Uri.parse(ApiEndpoint.loginUrl),
+      body: <String, String>{"email": email, "password": hashPassword},
+    );
+
+    final response = Response.fromJson(
+      raw.statusCode,
+      jsonDecode(raw.body) as Map<String, dynamic>,
+    );
+
+    return response;
+  }
+
+  Future<Response> register(String username, String email, String password) async {
+    final bytePassword = utf8.encode(password);
+    final hashPassword = sha256.convert(bytePassword).toString();
+
+    final raw = await http.post(
+      Uri.parse(ApiEndpoint.registerUrl),
+      headers: {'Content-Type': 'application/x-www-form-urlencoded'},
+      body: <String, String>{
+        "username": username,
+        "email": email,
+        "password": hashPassword,
+      },
+    );
+
+    final response = Response.fromJson(
+      raw.statusCode,
+      jsonDecode(raw.body) as Map<String, dynamic>,
+    );
+
+    return response;
   }
   
-  // Register new user
-  Future<UserModel?> register(String email, String password, String name) async {
-    try {
-      final response = await _client.post(
-        Uri.parse('$baseUrl/auth/register'),
-        headers: {'Content-Type': 'application/json'},
-        body: json.encode({
-          'email': email,
-          'password': password,
-          'name': name,
-        }),
-      );
-      
-      if (response.statusCode == 200) {
-        final jsonData = json.decode(response.body);
-        return UserModel.fromJson(jsonData['user']);
-      } else {
-        debugPrint('Registration failed: ${response.statusCode}');
-        return null;
-      }
-    } catch (e) {
-      debugPrint('Registration error: $e');
-      return null;
-    }
+  Future<Response> getProfile(String token, String userId) async {
+    final raw = await _client.get(
+      Uri.parse('${ApiEndpoint.userUrl}/$userId'),
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer $token',
+      },
+    );
+
+    final response = Response.fromJson(
+      raw.statusCode,
+      jsonDecode(raw.body) as Map<String, dynamic>,
+    );
+
+    return response;
   }
-  
-  // Get user profile
-  Future<UserModel?> getProfile(String token) async {
+    
+  // Verify token is still valid
+  Future<bool> verifyToken(String token) async {
     try {
+      // You'll need to add a token verification endpoint on your server
       final response = await _client.get(
-        Uri.parse('$baseUrl/auth/profile'),
+        Uri.parse('${ApiEndpoint.baseUrl}/verify-token'),
         headers: {
-          'Content-Type': 'application/json',
+          'Content-Type': 'application/x-www-form-urlencoded', 
           'Authorization': 'Bearer $token',
         },
       );
       
       if (response.statusCode == 200) {
         final jsonData = json.decode(response.body);
-        return UserModel.fromJson(jsonData);
-      } else {
-        debugPrint('Get profile failed: ${response.statusCode}');
-        return null;
+        return jsonData['code'] == 200;
       }
+      return false;
     } catch (e) {
-      debugPrint('Get profile error: $e');
-      return null;
+      debugPrint('Token verification error: $e');
+      return false;
     }
   }
 }
