@@ -7,7 +7,7 @@ part 'category_model.g.dart';
 @freezed
 abstract class CategoryModel with _$CategoryModel {
   const factory CategoryModel({
-    required String id,
+    required String categoryId,
     required String name,
     required bool isSensitive,
   }) = _CategoryModel;
