@@ -8,14 +8,14 @@ part of 'category_model.dart';
 
 _CategoryModel _$CategoryModelFromJson(Map<String, dynamic> json) =>
     _CategoryModel(
-      id: json['id'] as String,
+      categoryId: json['categoryId'] as String,
       name: json['name'] as String,
       isSensitive: json['isSensitive'] as bool,
     );
 
 Map<String, dynamic> _$CategoryModelToJson(_CategoryModel instance) =>
     <String, dynamic>{
-      'id': instance.id,
+      'categoryId': instance.categoryId,
       'name': instance.name,
       'isSensitive': instance.isSensitive,
     };
