@@ -1,4 +1,7 @@
+import 'package:flutter/foundation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:gameverse/domain/models/category_model/category_model.dart';
+import 'package:gameverse/domain/models/resource_model/resource_model.dart';
 
 part 'game_model.freezed.dart';
 part 'game_model.g.dart';
@@ -7,21 +10,27 @@ part 'game_model.g.dart';
 abstract class GameModel with _$GameModel {
   const factory GameModel({
     required String gameId,
+    required String publisherId,
     required String name,
     required int recommended,
     required String briefDescription,
     required String description,
     required String requirements,
-    required String headerImage,
-    List<String>? screenshots,
-    Map<String, dynamic>? price,
-    required List<String> categoriesID,
+    required ResourceModel headerImage,
+    required double price,
+    required List<CategoryModel> categories,
+    required List<ResourceModel>? resources,
+    required DateTime releaseDate,
 
     // Sale related fields
     bool? isSale,
     double? discountPercent,
     DateTime? saleStartDate,
     DateTime? saleEndDate,
+
+    // Download related fields
+    List<String>? binaries,
+    List<String>? exes,
 
     // Field for User only
     @Default(false) bool isOwned,
