@@ -7,28 +7,36 @@ part of 'game_model.dart';
 // **************************************************************************
 
 _GameModel _$GameModelFromJson(Map<String, dynamic> json) => _GameModel(
-  gameId: json['gameid'] as String,
+  gameId: json['gameId'] as String,
+  publisherId: json['publisherId'] as String,
   name: json['name'] as String,
   recommended: (json['recommended'] as num).toInt(),
   briefDescription: json['briefDescription'] as String,
   description: json['description'] as String,
   requirements: json['requirements'] as String,
-  headerImage: json['headerImage'] as String,
-  screenshots:
-      (json['screenshots'] as List<dynamic>?)?.map((e) => e as String).toList(),
-  price: json['price'] as Map<String, dynamic>?,
-  categoriesID:
-      (json['categoriesID'] as List<dynamic>).map((e) => e as String).toList(),
+  headerImage: ResourceModel.fromJson(
+    json['headerImage'] as Map<String, dynamic>,
+  ),
+  price: (json['price'] as num).toDouble(),
+  categories: (json['categories'] as List<dynamic>)
+      .map((e) => CategoryModel.fromJson(e as Map<String, dynamic>))
+      .toList(),
+  resources: (json['resources'] as List<dynamic>?)
+      ?.map((e) => ResourceModel.fromJson(e as Map<String, dynamic>))
+      .toList(),
+  releaseDate: DateTime.parse(json['releaseDate'] as String),
   isSale: json['isSale'] as bool?,
   discountPercent: (json['discountPercent'] as num?)?.toDouble(),
-  saleStartDate:
-      json['saleStartDate'] == null
-          ? null
-          : DateTime.parse(json['saleStartDate'] as String),
-  saleEndDate:
-      json['saleEndDate'] == null
-          ? null
-          : DateTime.parse(json['saleEndDate'] as String),
+  saleStartDate: json['saleStartDate'] == null
+      ? null
+      : DateTime.parse(json['saleStartDate'] as String),
+  saleEndDate: json['saleEndDate'] == null
+      ? null
+      : DateTime.parse(json['saleEndDate'] as String),
+  binaries: (json['binaries'] as List<dynamic>?)
+      ?.map((e) => e as String)
+      .toList(),
+  exes: (json['exes'] as List<dynamic>?)?.map((e) => e as String).toList(),
   isOwned: json['isOwned'] as bool? ?? false,
   installed: json['installed'] as bool? ?? false,
   favorite: json['favorite'] as bool? ?? false,
@@ -37,20 +45,24 @@ _GameModel _$GameModelFromJson(Map<String, dynamic> json) => _GameModel(
 
 Map<String, dynamic> _$GameModelToJson(_GameModel instance) =>
     <String, dynamic>{
-      'appId': instance.gameId,
+      'gameId': instance.gameId,
+      'publisherId': instance.publisherId,
       'name': instance.name,
       'recommended': instance.recommended,
       'briefDescription': instance.briefDescription,
       'description': instance.description,
       'requirements': instance.requirements,
       'headerImage': instance.headerImage,
-      'screenshots': instance.screenshots,
       'price': instance.price,
-      'categoriesID': instance.categoriesID,
+      'categories': instance.categories,
+      'resources': instance.resources,
+      'releaseDate': instance.releaseDate.toIso8601String(),
       'isSale': instance.isSale,
       'discountPercent': instance.discountPercent,
       'saleStartDate': instance.saleStartDate?.toIso8601String(),
       'saleEndDate': instance.saleEndDate?.toIso8601String(),
+      'binaries': instance.binaries,
+      'exes': instance.exes,
       'isOwned': instance.isOwned,
       'installed': instance.installed,
       'favorite': instance.favorite,
