@@ -1,4 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:gameverse/domain/models/resource_model/resource_model.dart';
 
 part 'user_model.freezed.dart';
 
@@ -8,12 +9,13 @@ part 'user_model.g.dart';
 abstract class UserModel with _$UserModel {
   const factory UserModel({
     required String id,
-    required String name,
+    required String username,
     required String email,
-    String? avatarUrl,
+    ResourceModel? avatar,
     String? bio,
     String? typeUser,
 
+    // Optional fields for user preferences
     List<String>? ownedGamesID,
     List<String>? favoriteGamesID,
     List<String>? installedGamesID,
