@@ -1,8 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:gameverse/domain/models/game_model/game_model.dart';
+import 'package:gameverse/config/spacing_config.dart';
+// import 'package:gameverse/domain/models/game_model/game_model.dart';
 import 'package:provider/provider.dart';
-import 'package:go_router/go_router.dart';
+// import 'package:go_router/go_router.dart';
 import 'package:gameverse/ui/forums/view_model/forum_viewmodel.dart';
+
+import 'package:gameverse/config/app_theme.dart';
+
+import 'game_forum_card.dart';
 
 class ForumsScreen extends StatefulWidget {
   const ForumsScreen({super.key});
@@ -41,297 +46,172 @@ class _ForumsScreenState extends State<ForumsScreen> {
             .toList();
 
         return SingleChildScrollView(
-          padding: const EdgeInsets.all(16),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              // Header
-              Row(
-                children: [
-                  Icon(
-                    Icons.forum,
-                    color: theme.colorScheme.primary,
-                    size: 28,
-                  ),
-                  const SizedBox(width: 12),
-                  Text(
-                    'Game Forums',
-                    style: theme.textTheme.headlineMedium?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Discuss your favorite games with the community',
-                style: theme.textTheme.bodyLarge?.copyWith(
-                  color: theme.colorScheme.onSurfaceVariant,
+          child: Padding(
+            padding: getNegativeSpacePadding(context),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                const SizedBox(height: 64),
+                // Header
+                Text(
+                  'Game Forums',
+                  style: theme.textTheme.displayLarge,
                 ),
-              ),
-              const SizedBox(height: 24),
-
-              // Search Bar
-              Container(
-                decoration: BoxDecoration(
-                  color: theme.colorScheme.surfaceContainerHighest,
-                  borderRadius: BorderRadius.circular(12),
-                  border: Border.all(
-                    color: theme.colorScheme.outline.withValues(alpha: 0.3),
+                const SizedBox(height: 8),
+                Text(
+                  'Discuss your favorite games with the community',
+                  style: theme.textTheme.bodyLarge?.copyWith(
+                    color: theme.colorScheme.onSurfaceVariant,
                   ),
                 ),
-                child: TextField(
-                  controller: _searchController,
-                  onChanged: (value) {
-                    setState(() {
-                      _searchQuery = value;
-                    });
-                  },
-                  decoration: InputDecoration(
-                    hintText: 'Search game forums...',
-                    prefixIcon: Icon(
-                      Icons.search,
-                      color: theme.colorScheme.onSurfaceVariant,
-                    ),
-                    suffixIcon: _searchQuery.isNotEmpty
-                        ? IconButton(
-                            icon: Icon(
-                              Icons.clear,
-                              color: theme.colorScheme.onSurfaceVariant,
-                            ),
-                            onPressed: () {
-                              _searchController.clear();
-                              setState(() {
-                                _searchQuery = '';
-                              });
-                            },
-                          )
-                        : null,
-                    border: InputBorder.none,
-                    contentPadding: const EdgeInsets.symmetric(
-                      horizontal: 16,
-                      vertical: 16,
-                    ),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 24),
-
-              // Results count
-              if (_searchQuery.isNotEmpty)
-                Padding(
-                  padding: const EdgeInsets.only(bottom: 16),
-                  child: Text(
-                    '${filteredGames.length} forum${filteredGames.length == 1 ? '' : 's'} found',
-                    style: theme.textTheme.bodyMedium?.copyWith(
-                      color: theme.colorScheme.onSurfaceVariant,
-                    ),
-                  ),
-                ),
-
-              // Content
-              if (viewModel.state == ForumsState.loading)
-                const Center(
-                  child: Padding(
-                    padding: EdgeInsets.all(32),
-                    child: CircularProgressIndicator(),
-                  ),
-                )
-              else if (viewModel.state == ForumsState.error)
-                Center(
-                  child: Column(
-                    children: [
-                      Icon(
-                        Icons.error_outline,
-                        size: 64,
-                        color: theme.colorScheme.error,
+                const SizedBox(height: 24),
+            
+                // Search Bar
+                SizedBox(
+                  // decoration: BoxDecoration(
+                  //   color: theme.colorScheme.surfaceContainerHighest,
+                  //   borderRadius: BorderRadius.circular(12),
+                  //   border: Border.all(
+                  //     color: theme.colorScheme.outline.withValues(alpha: 0.3),
+                  //   ),
+                  // ),
+                  child: TextField(
+                    controller: _searchController,
+                    onChanged: (value) {
+                      setState(() {
+                        _searchQuery = value;
+                      });
+                    },
+                    decoration: InputDecoration(
+                      hintText: 'Search game forums...',
+                      prefixIcon: Icon(
+                        Icons.search,
+                        color: AppTheme.currentThemeColors(theme.brightness).getText
                       ),
-                      const SizedBox(height: 16),
-                      Text(
-                        viewModel.errorMessage,
-                        style: theme.textTheme.titleMedium,
+                      suffixIcon: _searchQuery.isNotEmpty
+                          ? IconButton(
+                              icon: Icon(
+                                Icons.clear,
+                                color: theme.colorScheme.onSurfaceVariant,
+                              ),
+                              onPressed: () {
+                                _searchController.clear();
+                                setState(() {
+                                  _searchQuery = '';
+                                });
+                              },
+                            )
+                          : null,
+                      border: InputBorder.none,
+                      enabledBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: AppTheme.currentThemeColors(theme.brightness).getText),
+                        borderRadius: BorderRadius.zero
                       ),
-                      const SizedBox(height: 16),
-                      ElevatedButton(
-                        onPressed: () => viewModel.loadGamesWithForums(),
-                        child: const Text('Retry'),
+                      focusedBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: AppTheme.currentThemeColors(theme.brightness).getCyan),
+                        borderRadius: BorderRadius.zero
                       ),
-                    ],
+                      filled: false,
+                      contentPadding: const EdgeInsets.symmetric(
+                        horizontal: 16,
+                        vertical: 16,
+                      ),
+                    ),
                   ),
-                )
-              else if (filteredGames.isEmpty)
-                Center(
-                  child: Column(
-                    children: [
-                      Icon(
-                        _searchQuery.isNotEmpty ? Icons.search_off : Icons.forum_outlined,
-                        size: 64,
+                ),
+                const SizedBox(height: 24),
+            
+                // Results count
+                if (_searchQuery.isNotEmpty)
+                  Padding(
+                    padding: const EdgeInsets.only(bottom: 16),
+                    child: Text(
+                      '${filteredGames.length} forum${filteredGames.length == 1 ? '' : 's'} found',
+                      style: theme.textTheme.bodyMedium?.copyWith(
                         color: theme.colorScheme.onSurfaceVariant,
                       ),
-                      const SizedBox(height: 16),
-                      Text(
-                        _searchQuery.isNotEmpty 
-                            ? 'No forums found for "$_searchQuery"'
-                            : 'No forums available',
-                        style: theme.textTheme.titleMedium,
-                      ),
-                    ],
+                    ),
                   ),
-                )
-              else
-                // Column layout for forums
-                Column(
-                  children: filteredGames.map((game) {
-                    return Padding(
-                      padding: const EdgeInsets.only(bottom: 16),
-                      child: _GameForumCard(game: game),
-                    );
-                  }).toList(),
-                ),
-            ],
+            
+                // Content
+                if (viewModel.state == ForumsState.loading)
+                  const Center(
+                    child: Padding(
+                      padding: EdgeInsets.all(32),
+                      child: CircularProgressIndicator(),
+                    ),
+                  )
+                else if (viewModel.state == ForumsState.error)
+                  Center(
+                    child: Column(
+                      children: [
+                        Icon(
+                          Icons.error_outline,
+                          size: 64,
+                          color: theme.colorScheme.error,
+                        ),
+                        const SizedBox(height: 16),
+                        Text(
+                          viewModel.errorMessage,
+                          style: theme.textTheme.titleMedium,
+                        ),
+                        const SizedBox(height: 16),
+                        ElevatedButton(
+                          onPressed: () => viewModel.loadGamesWithForums(),
+                          child: const Text('Retry'),
+                        ),
+                      ],
+                    ),
+                  )
+                else if (filteredGames.isEmpty)
+                  Center(
+                    child: Column(
+                      children: [
+                        Icon(
+                          _searchQuery.isNotEmpty ? Icons.search_off : Icons.forum_outlined,
+                          size: 64,
+                          color: theme.colorScheme.onSurfaceVariant,
+                        ),
+                        const SizedBox(height: 16),
+                        Text(
+                          _searchQuery.isNotEmpty 
+                              ? 'No forums found for "$_searchQuery"'
+                              : 'No forums available',
+                          style: theme.textTheme.titleMedium,
+                        ),
+                      ],
+                    ),
+                  )
+                else
+                  // Column layout for forums
+                  // Column(
+                  //   children: [
+                  //     ListView.separated(
+                  //       physics: NeverScrollableScrollPhysics(),
+                  //       scrollDirection: Axis.vertical,
+                  //       itemCount: filteredGames.length,
+                  //       itemBuilder: (context, index) {
+                  //         return GameForumCard(game: filteredGames[index]);
+                  //       },
+                  //       separatorBuilder:(context, index) => const SizedBox(height: 16),
+                  //     ),
+                  //   ],
+                  // ),
+                  Column(
+                    children: filteredGames.map((game) {
+                      return Column(
+                        children: [
+                          GameForumCard(game: game),
+                          const SizedBox(height: 16)
+                        ],
+                      );
+                    }).toList(),
+                  ),
+              ],
+            ),
           ),
         );
       },
-    );
-  }
-}
-
-class _GameForumCard extends StatelessWidget {
-  final GameModel game;
-
-  const _GameForumCard({required this.game});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return Card(
-      clipBehavior: Clip.antiAlias,
-      elevation: 2,
-      child: InkWell(
-        onTap: () {
-          context.push('/forum-posts/${game.gameId}/${Uri.encodeComponent(game.name)}');
-        },
-        child: SizedBox(
-          height: 120,
-          child: Row(
-            children: [
-              // Game Image with gradient opacity
-              Expanded(
-                flex: 2,
-                child: Stack(
-                  children: [
-                    SizedBox(
-                      width: double.infinity,
-                      height: double.infinity,
-                      child: Image.network(
-                        game.headerImage,
-                        fit: BoxFit.cover,
-                        errorBuilder: (context, error, stackTrace) {
-                          return Container(
-                            color: theme.colorScheme.surfaceContainerHighest,
-                            child: const Center(
-                              child: Icon(Icons.videogame_asset, size: 32),
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                    // Gradient overlay from right to left (opacity 1 to 0)
-                    Container(
-                      decoration: BoxDecoration(
-                        gradient: LinearGradient(
-                          begin: Alignment.centerRight,
-                          end: Alignment.centerLeft,
-                          colors: [
-                            Colors.black.withValues(alpha: 0.7),
-                            Colors.black.withValues(alpha: 0.3),
-                            Colors.transparent,
-                          ],
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              
-              // Game Info
-              Expanded(
-                flex: 3,
-                child: Padding(
-                  padding: const EdgeInsets.all(16),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            game.name,
-                            style: theme.textTheme.titleLarge?.copyWith(
-                              fontWeight: FontWeight.bold,
-                            ),
-                            maxLines: 2,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                          const SizedBox(height: 4),
-                          Text(
-                            game.briefDescription,
-                            style: theme.textTheme.bodyMedium?.copyWith(
-                              color: theme.colorScheme.onSurfaceVariant,
-                            ),
-                            maxLines: 2,
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ],
-                      ),
-                      Row(
-                        children: [
-                          Container(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 12,
-                              vertical: 6,
-                            ),
-                            decoration: BoxDecoration(
-                              color: theme.colorScheme.primaryContainer,
-                              borderRadius: BorderRadius.circular(16),
-                            ),
-                            child: Row(
-                              children: [
-                                Icon(
-                                  Icons.chat_bubble_outline,
-                                  size: 16,
-                                  color: theme.colorScheme.onPrimaryContainer,
-                                ),
-                                const SizedBox(width: 4),
-                                Text(
-                                  'Join Discussion',
-                                  style: TextStyle(
-                                    color: theme.colorScheme.onPrimaryContainer,
-                                    fontWeight: FontWeight.w500,
-                                    fontSize: 12,
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                          const Spacer(),
-                          Icon(
-                            Icons.arrow_forward_ios,
-                            size: 16,
-                            color: theme.colorScheme.onSurfaceVariant,
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
