@@ -29,6 +29,7 @@ func getUser(c echo.Context, client *supabase.Client) error {
 	if err != nil {
 		return jsonResponse(c, http.StatusBadRequest, "Invalid userid" /*err.Error()*/, "")
 	}
+	user["id"] = userid
 	return jsonResponse(c, http.StatusOK, "", user)
 }
 
@@ -45,6 +46,7 @@ func addUser(c echo.Context, client *supabase.Client) error {
 	}
 	_, _, err := client.From("User").Insert(data, false, "", "", "").ExecuteString()
 	if err != nil {
+		fmt.Println("Error inserting user:", err)
 		return jsonResponse(c, http.StatusBadRequest, err.Error(), "")
 	}
 
@@ -125,12 +127,12 @@ func login(c echo.Context, client *supabase.Client) error {
 	if err != nil {
 		return jsonResponse(c, http.StatusBadRequest, err.Error(), "")
 	}
-
-	returnVal := map[string]string{
+	// return jsonResponse(c, http.StatusOK, "", createUserToken(userid["userid"]))
+	// Return the user ID and token
+	return jsonResponse(c, http.StatusOK, "", map[string]string{
 		"userid": userid["userid"],
 		"token":  createUserToken(userid["userid"]),
-	}
-	return jsonResponse(c, http.StatusOK, "", returnVal)
+	})
 }
 
 func getGamesWithStatus(c echo.Context, client *supabase.Client, userid string, status string) error {
